requirements.txt
easydict~=1.9
pyyaml~=6.0
numpy~=1.21.5
pillow~=9.0.1
scipy~=1.7.3
pytorch-model-summary~=0.1.2
matplotlib~=3.5.1
p_tqdm~=1.3.3
tqdm~=4.63.0
torchvision~=0.12.0
torch~=1.11.0
--extra-index-url https://download.pytorch.org/whl/cu113
